toolkit/components/passwordmgr/test/test_basic_form_0pw.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Password Manager</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>  
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Password Manager test: forms with no password fields
<p id="display"></p>

<div id="content" style="display: none">

  <!-- Form with no user field or password field -->
  <form action="formtest.js" method="get">
    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <!-- Form with no user field or password field, but one other field -->
  <form action="formtest.js" method="get">
    <input type="checkbox">
    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <!-- Form user a user field, but no password field -->
  <form action="formtest.js" method="get">
    <input  type="text"       name="uname1" value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

</div>

<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Password Manger: form fill, no password fields. **/

function startTest() {
  is($_("uname1").value, "", "Checking for unfilled username 1");

  SimpleTest.finish();
}

window.onload = startTest;

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
